requirements.txt
fastapi
pandas
uvicorn
openpyxl
xlrd
PyYAML
beautifulsoup4
